Fix combiner.py for newer version of hail

+++ combiner.py
@@ -3,9 +3,10 @@ import hail as hl
 
 from itertools import islice
 from argparse import ArgumentParser
-from os import getcwd, makedirs, environ
-from os.path import join, realpath, expanduser
+from os import makedirs, environ
+from os.path import exists, expanduser, join, realpath
 from shutil import rmtree
+import logging
 
 from pyspark import SparkConf, SparkContext
 
@@ -13,6 +14,8 @@ if __name__ == '__main__':
     parser = ArgumentParser()
     parser.add_argument(metavar='input-file', dest='path_to_input_list',
                         help='a file that lists VCFs (i.e., `*.vcf` or `*.vcf.gz`)')
+    parser.add_argument('--vds', type=str, action='append',
+                        help='a path to the VDS directory')
     parser.add_argument('-n', type=int,
                         help='only use first n items from the input file (default: unlimited)')
     parser.add_argument('--master', dest='master', type=str,
@@ -35,17 +38,25 @@ if __name__ == '__main__':
                         help='#cores for each executor')
     parser.add_argument('--driver-cores', dest='driver_cores', type=str,
                         help='#cores for each driver')
-    parser.add_argument('--branch-factor', dest='branch_factor', type=int,
+    parser.add_argument('--branch-factor', dest='branch_factor', type=int, default=100,
                         help='branch factor to hierarchically merge GVCFs')
-    parser.add_argument('--batch-size', dest='batch_size', type=int,
-                        help='batch size')
+    # parser.add_argument('--batch-size', dest='batch_size', type=int,
+    #                     help='batch size')
     # parser.add_argument('--force', help='force output to the existing directory')
     args = parser.parse_args()
+
+    if args.vds is None:
+        input_vdses = None
+    else:
+        input_vdses = [realpath(expanduser(v)) for v in args.vds]
+
+    FORMAT = '%(asctime)-15s: %(message)s'
+    logging.basicConfig(level=logging.INFO, format=FORMAT)
 
     conf = {
         'spark.executor.memory': args.executor_memory,
         'spark.driver.memory': args.driver_memory,
-        'spark.yarn.am.memory': args.driver_memory,
+        # 'spark.yarn.am.memory': args.driver_memory,
         # 'spark.jars': join(environ['HAIL_DIR'], 'hail-all-spark.jar'),
         # 'spark.execuor.extraClassPath': './hail-all-spark.jar',
         # 'spark.driver.extraClassPath': join(environ['HAIL_DIR'], 'hail-all-spark.jar'),
@@ -57,7 +68,7 @@ if __name__ == '__main__':
         conf['spark.executor.cores'] = str(args.executor_cores)
     if args.driver_cores is not None:
         conf['spark.driver.cores'] = str(args.driver_cores)
-        conf['spark.yarn.am.cores'] = str(args.driver_cores)
+        # conf['spark.yarn.am.cores'] = str(args.driver_cores)
 
     if args.master is None:
         sc = None
@@ -66,7 +77,7 @@ if __name__ == '__main__':
         con = SparkConf()
         con.setMaster(args.master)
         con.set('spark.submit.deployMode', 'client')
-        con.set('spark.yarn.stagingDir', 'lustre:///lustre8/home/tanjo-pg')
+        # con.set('spark.yarn.stagingDir', 'lustre:///lustre8/home/tanjo-pg')
         con.set('spark.jars', join(environ['HAIL_DIR'], 'hail-all-spark.jar'))
         # con.set('spark.files', join(environ['HAIL_DIR'], 'hail-all-spark.jar'))
         # con.set('spark.execuor.extraLibraryPath', '/home/tanjo-pg/miniconda3/envs/hail-env/lib')
@@ -84,37 +95,50 @@ if __name__ == '__main__':
     # else:
     #     base_tmpdir = realpath(args.tmpdir)
 
-    base_tmpdir = realpath(expanduser('~/.tmp'))
+    out_base = realpath(expanduser(args.out_base))
+    makedirs(out_base)
+    logfile = join(out_base, 'output.log')
+
+    base_tmpdir = join(out_base, 'tmp')
     network_tmpdir = join(base_tmpdir, 'network-tmp')
     makedirs(network_tmpdir)
     combiner_tmpdir = join(base_tmpdir, 'combiner-tmp')
     makedirs(combiner_tmpdir)
 
-    out_base = realpath(args.out_base)
-    makedirs(out_base)
-    logfile = join(out_base, 'output.log')
-    output = join(out_base, 'output.mt')
-
     try:
         inputs = []
         hl.init(sc=sc, spark_conf=spark_conf, log=logfile, tmp_dir=network_tmpdir)
-        with hl.hadoop_open(realpath(args.path_to_input_list), 'r') as f:
+        with open(realpath(expanduser(args.path_to_input_list)), 'r') as f:
             if args.n is None:
                 inputs = [line.strip() for line in f]
             else:
                 inputs = list(islice((line.strip() for line in f), 0, args.n))
 
-        if args.branch_factor:
-            factor = args.branch_factor
-        else:
-            factor = 100
+        output = join(out_base, 'output.vds')
+        rg = hl.get_reference('GRCh38')
+        combiner = hl.vds.new_combiner(
+            output_path=output,
+            temp_path=combiner_tmpdir,
+            gvcf_paths=inputs,
+            vds_paths=input_vdses,
+            use_genome_default_intervals=True,
+            reference_genome=rg,
+            branch_factor=args.branch_factor,
+        )
 
-        if args.batch_size:
-            bsize = args.batch_size
-        else:
-            bsize = 100
-        hl.experimental.run_combiner(inputs, out_file=output, tmp_path=combiner_tmpdir,
-                                     use_genome_default_intervals=True, branch_factor=factor, batch_size=bsize,
-                                     reference_genome='GRCh38')
+        combiner.run()
+
+        logging.info("Combiner completed")
+        vds = hl.vds.read_vds(output)
+        logging.info("Exporting VCF...")
+        hl.methods.export_vcf(vds, join(out_base, 'output.vcf.bgz'))
+        logging.info("Exporting VCF...: done!")
+    except Exception as e:
+        logging.warning(f"Caught exception: {type(e).__name__}({e})")
+        if exists(out_base):
+            logging.info(f"In exception: remove {out_base}")
+            rmtree(out_base)
     finally:
-        rmtree(base_tmpdir)
+        if exists(base_tmpdir):
+            logging.info(f"Clean up: remove {base_tmpdir}")
+            rmtree(base_tmpdir)
